+++ datasetinsights/datasets/synthetic.py
@@ -8,13 +8,13 @@ from PIL import Image
 from sklearn.model_selection import train_test_split
 
 import datasetinsights.constants as const
-from datasetinsights.data.bbox import BBox2D
-from datasetinsights.datasets.simulation import AnnotationDefinitions, Captures
-from datasetinsights.datasets.simulation.download import (
-    Downloader,
-    download_manifest,
+from datasetinsights.datasets.unity_perception import (
+    AnnotationDefinitions,
+    Captures,
 )
-from datasetinsights.datasets.simulation.tables import SCHEMA_VERSION
+from datasetinsights.datasets.unity_perception.tables import SCHEMA_VERSION
+from datasetinsights.io.bbox import BBox2D
+from datasetinsights.io.usim import Downloader, download_manifest
 
 from .base import Dataset
