+++ datasetinsights/estimators/densedepth.py
@@ -15,10 +15,10 @@ from torchvision.transforms import Compose
 from torchvision.transforms.functional import to_tensor
 
 import datasetinsights.constants as const
-from datasetinsights.data.loader import create_loader
-from datasetinsights.data.transforms import RandomHorizontalFlip, Resize
 from datasetinsights.datasets import Dataset
 from datasetinsights.evaluation_metrics import EvaluationMetric
+from datasetinsights.io.loader import create_loader
+from datasetinsights.io.transforms import RandomHorizontalFlip, Resize
 from datasetinsights.stats.visualization.plots import grid_plot
 
 from .base import Estimator
