+++ datasetinsights/io/usim.py
@@ -12,10 +12,12 @@ from requests.packages.urllib3.util.retry import Retry
 from tqdm import tqdm
 
 import datasetinsights.constants as const
-from datasetinsights.data.download import TimeoutHTTPAdapter, download_file
-from datasetinsights.data.exceptions import DownloadError
-
-from .tables import DATASET_TABLES, FileType
+from datasetinsights.datasets.unity_perception.tables import (
+    DATASET_TABLES,
+    FileType,
+)
+from datasetinsights.io.download import TimeoutHTTPAdapter, download_file
+from datasetinsights.io.exceptions import DownloadError
 
 logger = logging.getLogger(__name__)
 # number of workers for ThreadPoolExecutor. This is the default value
