+++ datasetinsights/storage/checkpoint.py
@@ -5,7 +5,7 @@ import os
 import tempfile
 
 import datasetinsights.constants as const
-from datasetinsights.data.download import download_file
+from datasetinsights.io.download import download_file
 from datasetinsights.torch_distributed import is_master
 
 from .gcs import GCSClient, gcs_bucket_and_path
