+++ tests/test_kitti.py
@@ -4,12 +4,12 @@ import numpy as np
 import torch
 from pyquaternion import Quaternion
 
-from datasetinsights.data.bbox import BBox3d
 from datasetinsights.datasets.kitti import (
     KittiBox3d,
     KittiTransforms,
     convert_kitti2nu,
 )
+from datasetinsights.io.bbox import BBox3d
 
 CUR_FILE = os.path.dirname(os.path.abspath(__file__))
