+++ tests/unity_perception/test_captures.py
@@ -3,9 +3,14 @@ import json
 
 import pytest
 
-from datasetinsights.datasets.simulation import Captures
-from datasetinsights.datasets.simulation.exceptions import DefinitionIDError
-from datasetinsights.datasets.simulation.tables import SCHEMA_VERSION, glob
+from datasetinsights.datasets.unity_perception import Captures
+from datasetinsights.datasets.unity_perception.exceptions import (
+    DefinitionIDError,
+)
+from datasetinsights.datasets.unity_perception.tables import (
+    SCHEMA_VERSION,
+    glob,
+)
 
 
 def test_get_captures_and_annotations(mock_data_dir):
