+++ tests/unity_perception/test_metrics.py
@@ -4,9 +4,14 @@ import json
 import pandas as pd
 import pytest
 
-from datasetinsights.datasets.simulation import Metrics
-from datasetinsights.datasets.simulation.exceptions import DefinitionIDError
-from datasetinsights.datasets.simulation.tables import SCHEMA_VERSION, glob
+from datasetinsights.datasets.unity_perception import Metrics
+from datasetinsights.datasets.unity_perception.exceptions import (
+    DefinitionIDError,
+)
+from datasetinsights.datasets.unity_perception.tables import (
+    SCHEMA_VERSION,
+    glob,
+)
 
 
 def test_filter_metrics(mock_data_dir):
