+++ tests/unity_perception/test_references.py
@@ -1,10 +1,13 @@
 import json
 
-from datasetinsights.datasets.simulation import (
+from datasetinsights.datasets.unity_perception import (
     AnnotationDefinitions,
     MetricDefinitions,
 )
-from datasetinsights.datasets.simulation.tables import SCHEMA_VERSION, glob
+from datasetinsights.datasets.unity_perception.tables import (
+    SCHEMA_VERSION,
+    glob,
+)
 
 
 def test_annotation_definitions(mock_data_dir):
